backend/pyproject.toml
[tool.poetry]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = ["Tim Chu <tchu@indeed.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
strawberry-graphql = "^0.278.0"
fastapi = {extras = ["standard"], version = "^0.116.1"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
